requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.2
aiohttp==3.9.1
cryptography==41.0.7
flask==3.0.0
flask-cors==4.0.0